Added greedy solver
Added a function that looks for the best move, based on a greedy
approach. The greedy algorithm minimizes the number of groups left after
the move is made.

Also changed the Grid class. It now stores info on the groups that have
been calculated, which in turn allowed the rewriting of the functions
findPossibleMoves and isFinished.